+++ CHANGELOG.md
@@ -5,8 +5,10 @@ changed functionality, fixed bugs).
 
 ## Version 2.1.1 (in development)
 
-- Always log the version of Flamenco Manager.
+- Always log the version of Flamenco Worker.
 - Include missing merge-exr.blend, required for progressive rendering, in the distribution bundle.
+- Include `exr-merge` task type in default configuration, which is required for progressive
+  rendering.
 
 
 ## Version 2.1.0 (2018-01-04)